test: refactor chiplets aux trace tests
